Use union_field_ref instead of set_* in union deserialization
Summary: There is a bug in union field with cpp.ref in setter. Regardless of the bug, we should try to use `union_field_ref` instead of using setters. Note, this diff does not fix the bug in union field with cpp.ref in setter. That will be handled separately.

Reviewed By: vitaut

Differential Revision: D49977738

fbshipit-source-id: 285016879a6669ae4724d238c8df700358b3b75b