win32: increment recrawlCount on recrawl
Summary:
Our telemetry appears to be showing that no sync_to_now request on Windows are
timing out, which feels surprising since we would expect queries being serviced
right after a recrawl to be timing out. To understand if recrawls are actually
happening, let's increment the recrawlCount counter which is being made
available in telemetry.

Reviewed By: genevievehelsel

Differential Revision: D50085940

fbshipit-source-id: 3c0b7e920bf74e28f8292c3687d12f89a38d3528