Improve IDL reference for constants
Summary:
The term "constant" is too ambiguous in the IDL reference. It can refer to a constant definition, a target representation of a constant (object) or a constant value. Make it less ambiguous by replacing "constant" with "initializer" to denote the constant initializer. This term is commonly used in the C family of languages (see e.g. https://docs.oracle.com/javase/specs/jls/se7/html/jls-10.html#jls-10.6) and is somewhat better because the construct is not a constant per se and doesn't even have an exact type but something used to initialize a constant or a field.

The issue is particularly obvious for the perviously used term "list constant" which may denote a constant (definition) of a list type or a list "value":

```
const list<i32> c = ... // list constant
... [1, 2, 3] // list constant but of a different kind
```

Clarify that initializers may not have exact types on their own and that the types are inferred. Inference rules will be defined in a separate diff.

Remove excessive section nesting.

Fix a typo: contant -> constant.

Reviewed By: aristidisp

Differential Revision: D50084597

fbshipit-source-id: 7b7e558e78a753a506cef40e3da693fefbfb1570