add a Retry-After header when returning 503 due to queue timeouts & shutdown
Summary:
We want to signal to clients that a given HHVM instance is overloaded, and they should retry. In the presence of reverse proxies, clients will likely hit another instance; in the absence of reverse proxies, this is still beneficial as the host may be functional after the interval.

`-1` is used as a sentinel value to disable the behaviour as MDN says `delay-seconds` is non-negative: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Retry-After

We do not support `http-date` style `Retry-After` here for simplicity.

Reviewed By: binliu19

Differential Revision: D49954310

fbshipit-source-id: ee67ac101c9340fa4f1ba40471f3bc2caf22814c