Cache 1 and 2 char prefixes
Summary:
In this diff, I create a cache for 1 and 2 character letter combinations that we can check for autocomplete suggestions. I believe this would enable us to remove the if check for things like 'bk' getting suggestions despite being only 2 characters.

As expected, there is a big slowdown (500-800ms) for the first lookup (a cache miss) but by about the 4th lookup, it normalizes and its fast (around 100ms).

Big thanks to ljw1004 for the idea here, and nt591 and vassilmladenov for the pointers!

Differential Revision: D49845430

fbshipit-source-id: c569e7157df0a145eb88f9492ceb0293a6b7451b