Fix handling of qualified interaction names
Summary:
Using a qualified interaction name currently results in an error. For example:

```
include "foo.thrift"
service S {
  foo.I createI();
}
```

gives

```
[ERROR:foo.thrift:3] Invalid first response type: foo.I
```

Fix this by not qualifying already qualified identifiers when resolving interaction names. This should really be handled on the scope level but that's a bigger change which is out of scope of this quick fix.

Reviewed By: rmakheja

Differential Revision: D50086585

fbshipit-source-id: f03e7fc93c4f310f0f5faf8753ae787ddfffa13b