Do not pessimise this type in return
Summary:
Our implicit pessimisation policy is to add a like type to any return type that is not enforced *fully* at runtime. This includes the `this` type, because HHVM does not enforce generic arguments.

However, in practice it's almost always ok to avoid adding a like to `this`. For example, a very common pattern is direct or indirect use of `return $this;`.

Reviewed By: madgen

Differential Revision: D48904625

fbshipit-source-id: 461d1d2c090ad7eda00dfb5ea23722d2430086f4