Add backtrace to intercept logger
Summary:
I'm seeing functions that are allowlisted in the automocker, but appear in this scuba intercept dataset. I'm curious how this is possible and how we are registering an intercept handler for functions if we are not automocking them.

ie `Coeffects\backdoor` is here
https://fburl.com/scuba/hhvm_intercept_function/6t6nzf04
and here https://fburl.com/code/z073rx0z

Reviewed By: ricklavoie

Differential Revision: D50124491

fbshipit-source-id: aa75a206b3f88edd954fd5ae4ff4b1c554d7b673