Revert D50082549: Multisect successfully blamed "D50082549: Log fanout information with log category "fanout_information"" for test or build failures
Summary:
This diff is reverting D50082549
D50082549: Log fanout information with log category "fanout_information" by CatherineGasnier has been identified to be causing the following test or build failures:

Tests affected:
- [hphp/hack/test/fanout:fanout_saved_state - main](https://www.internalfb.com/intern/test/844425023351573/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3253224
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D50131008

fbshipit-source-id: b7a3d5e879144e0aea9652c14186625215f34054