Redo and fix tests for "D50082549: Log fanout information with log category "fanout_information"" for test or build failures"
Summary:
That diff was automatically reverted because it broke //hphp/hack/test/fanout:fanout_saved_state

The printed JSON in this diff needs to be single line for the test infra for that test to be able to process it.
We also rename the log category used by the tests to fanout_tests, so it's clearer what these log lines are for.

Reviewed By: nt591

Differential Revision: D50172279

fbshipit-source-id: 62de49e11f0ff86a06c1700bdc4f2b683e00eefa