Allow padding in SimpleJSONProtocol base64 encoded binary fields
Summary:
Before this diff, the base64 decoder incorrectly deserialized base64 encoded binary fields with `=` padding. It corrupted the trailing bytes. This diff adds support for deserializing base64 encoded binary fields with `=` padding, but doesn't change the serializer.

This is an inconsistency with the `thrift-py` `SimpleJSONProtocol` deserializer, which does support base64 encoded strings with padding.

Additionally, the `thrift-py` `SimpleJsonProtocol` serializer doesn't base64 encode binary fields itself. It requires the user to base64 the field manually. This is because the generated code calls `Protocol.writeString()` for binary fields, rather than `Protocol.writeBinary()`.

And every base64 implementation out there will pad the base64 with `=`. The exception is base64 URL encoding, but Thrift doesn't use that format.

So the Thrift decoder should be generous in what it accepts.

Reviewed By: vitaut

Differential Revision: D49749440

fbshipit-source-id: af3cd4b8d6f84c98170f62e33743d7c697974ddc