ban invocation of static methods on traits with `require class` constraints
Summary:
Banning static calls of the form `T::foo()` where `T` is a trait with a `require` class attribute, as these have been responsible for S368253.

At this time WWW is clean, but the new error code (4476) is fixmeable so new errors  - if any - won't block the next Hack release.

Reviewed By: vsiles

Differential Revision: D49770426

fbshipit-source-id: 5629769cb24cf918c47b9861fae8e11732e465f6