justknob enhancements to support bool and int keys
Summary:
hh_distc justknob support was the bare minimum to try and disable the trace store via JustKnob as a backup plan. We never used it. This diff adds more functional parity with the OCaml usage.

## Updates

Fetching an undefined Justknob will panic at runtime, as OCaml does.

I've added an Enum to support the two JK values, that I know of anyway

## Gaps

There are plenty of JK fields in serverLocalConfigKnobs.ml I didn't port over, because it's just a bit of drudgery that wouldn't add new features.

Differential Revision: D50176305

fbshipit-source-id: b013d9d2423f2aff9586b762096d879b12e378ac