Remove dead support for not resetting state across requests
Summary:
The code around needReset() creates an impression that we might not reset state
across xbox requests (which would be terribly wrong). However, we use every
instance of XboxRequestHandler only once due to the destroyRequestHandler()
call in XboxWorker (which is itself bad, as the point of these request handlers
is to handle multiple requests on the same thread).

So, kill the logic, kill the options and move xbox request shutdown logic to
the teardownRequest() handler where it belongs.

Reviewed By: ricklavoie

Differential Revision: D50249107

fbshipit-source-id: b52c6f596a41e15d97ddd4d503e172e772161900