Remove flag optimized_private_member_fanout and use optimized_member_fanout instead
Summary: With the legacy member fanout algo, when a private member X::m was modified, it didn't matter that the change was marked as 'private', because there would be no user of the private member outside the (non-trait) type, so no edge from X::m in the graph. With the new fanout algorithm, it matters, because we don't just rely on edges from X::m. Therefore the flag optimized_private_member_fanout is not a separate optimisation, it is necessary to not regress fanouts with the new algorithm. This diff merges the two flags optimized_private_member_fanout and optimized_member_fanout.

Differential Revision: D50125024

fbshipit-source-id: 54b3dd07f6ed603dfb4f4a5bacc00fd98a02f3ad