Optimized member fanout: do not add require extends deps to fanout
Summary:
Reminder of the optimised member fanout algo:
 {F1116709932}

So prior to this diff, the fanout includes all descendant X of C down to the first descendant redeclaring m, in addition to all users of X::m for those descendant X.
With this diff: we still include all users of X::m for the same set of descendants, but if descendant X descends via a `require extends/implements` statements, then it's not included in the fanout, since it does not need to be rechecked by `Typing_extends` itself.

This is to prevent a regression compared to the legacy algorithm, which for modified members, relied solely on edges from C::m (given that using X::m would add an edge from X::m + an edge from C::m where C the origin of m in X. Checking that X::m is a subtype of C::m in Typing_extends would also add an edge from C::m).

See added tests for some examples.

How do we distinguish between 'Extends' edges and 'require extends' edges? We add a new kind of edge 'RequireExtends'. So far, an 'Extends x' node can be derived from the 'Type x' node by resetting the LSB. For 'RequireExtends', there's no need to reserve an additional bit. Instead we do `hash2 KRequireExtends (make (Type x)) ""`, in other words we rehash the hash for 'Type x'. We honestly could have done the same for Extends edges, but it might be wiser to have a way to distinguish 'Type' nodes from other nodes by checking that the LSB  is 1 when we traverse the depgraph.

Differential Revision: D50081368

fbshipit-source-id: a00368d2a16a6b6ce745b6b3ab24d557a5f04293