Revert D48068587: Multisect successfully blamed "D48068587: [Zonk] True type adjustments" for test or build failures
Summary:
This diff is reverting D48068587
D48068587: [Zonk] True type adjustments by fahndrich has been identified to be causing the following test or build failures:

Tests affected:
- [hphp/hack/test/facebook/ai/scenario/vc_creation:vc_creation - main](https://www.internalfb.com/intern/test/281475038921037/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3288546
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D50290047

fbshipit-source-id: 6bdbe61eddeffee9d1ca162e095912cb1af50a7c