Move common fields from Function/Method into Func
Summary:
Even though they represent generic func data `attributes`, `attrs`, and `coeffects` are stored in HHBC on the `Function` and `Method` structs instead of on the `Func`.  The IR duplicated this structure but not for any good reason.

Move those common fields onto `Func` and update users to reflect the change:  Readers were easy since any code reading the fields already had the outer struct (`Function` or `Method`) and could just reference the inner `func` field.   Writers are a little more complex because we sometimes builder the `Func` separately from where we build the `Function`/`Method`.

Reviewed By: edwinsmith

Differential Revision: D50160930

fbshipit-source-id: 4141f6dc8d244c1966a8798a895d650585ba3c0f