Add IsTypeStructShallow opcode
Summary:
For match statements and typed locals, we would like to discriminate and enforce
types in ways which are not currently possible.

We would like enforcement of typed locals to be similar to enforcement of
function parameters, with none of the recursive enforcement we do for tuples and
shapes in as expressions. This would allow users to add type hints to locals
without needing to worry about introducing new exceptions in cases where Hack
types are incorrect, or introducing expensive implicit type tests on every
assignment of the local.

Additionally, when using a value of a case type in a match statement, shallow
type discrimination will be sufficient to distinguish between the case type
variants, since their data types may not overlap, so we'd like to avoid the
deeper type testing behavior of is expressions.

Both for typed locals and match statements, it would be ideal if we could
support type constant accesses like this::T, which are currently supported by
is/as and reified generic param verification (but only for reified generic
params--$t in function f(this::T $t) {} is not currently enforced). AIUI, this
behavior cannot be expressed with the bytecodes currently offered. We need
something with the shallow-enforcement behavior of VerifyParamTypeTS, but which
can be used in any position like IsTypeStructC (rather than only at the function
entrypoint), and which supports enforcement of type constants.

This diff revises the IsTypeStructC bytecode to include a TypeStructEnforceKind
immediate, specifying whether the type assertion should be deep (used in
`is`/`as` tests; the current behavior) or shallow (used in reified param
enforcement, typed local enforcement, and match statement branching). It also
adds a new IR opcode IsTypeStructShallow to implement this test in the JIT.

Reviewed By: aorenste

Differential Revision: D49202529

fbshipit-source-id: 9c185c45af376f075ba83d61f1e12848b35f5a99