Clean up packages and module references implementation, improve error messaging
Summary:
This diff rips out the module references implementation in the typechecker which is unused, and also improves error messaging for packages when a module is undefined.

If a module is not defined, we can still use the module name as if it were defined to check package errors.

This diff hopefully cleans up the packages implementation so it's more understandable, both for the user and for us.

Differential Revision: D50329658

fbshipit-source-id: 7452ca03971f3e6c65ebf2ed85ecc00308f93e6f