Kill unnecessary kPurgedGenCount state
Summary:
eraseImpl() holds a write lock on the StoreValue it is erasing. After it sets
the state to kPurgedGenCount, it always erases the whole entry, so this state
is not visible by anyone else. Do not bother setting it to the kPurgedGenCount
state and checking that state at other place.

Instead, assert that the state is kIdleGenCount. It must be kIdleGenCount, as
we checked that the expiration time was not the special value of 1 and the
expireRequestIdx is only used with this value.

Except that there was a race condition, when one thread in checkExpire() may
have entered the code path to bump TTL, while another thread may see the value
as expired, setting up expireRequestIdx and setting expire to 1. This may end
up in incorrect state, with changed expireRequestIdx, but bumped TTL. Fix this
by using compare_exchange_strong to not overwrite the expire set to 1.

Reviewed By: ricklavoie

Differential Revision: D50345342

fbshipit-source-id: a6fbcc36593b15287aa49b3d82ebea2b8b294433