Remove treadmill-bypass optimization that kicks in only incorrectly
Summary:
We call purgeExpired() at the end of every request, so with sufficient request
frequency, we would expire entries shortly after the start of every second.

By definition, there is at least one request that started at the same second
(the one running the purgeExpired()), so the `e >= oldestLive` condition will
be true, aborting the optimization.

Except of a bug in getOldestStartTime() that incorrectly bumps the timestamp
by one second, creating a possibility of immediately freeing a value in one
request that is still being used by another request.

This removes APC's dependence on any Treadmill internals.

Reviewed By: ricklavoie

Differential Revision: D50394863

fbshipit-source-id: 6e183ed3f79097f5dce37cd9859dde70ac9bdf6e