More ServerProgress.with_message, requiring refactoring
Summary:
For now the progress message "figuring out files to recheck" shows up for a long time.
I want to show it with `ServerProgress.with_message` instead of `ServerProgress.write` (see previous diff) to see if it really takes a long time to compute fanout or if it's just because we don't update the progress message once we're done.
For the code to be more readable, I also bread down `post_saved_state_initialization` into smaller functions.

Reviewed By: ljw1004

Differential Revision: D50344990

fbshipit-source-id: b0283697ac17c8984864961fda6a36cf047717c9