Reason printing: simplify lazy string printing
Summary:
I thought it would be useful if the pretty printer of lazy strings in reasons shows whether the string was evaluated or not, but the hashing of lazy string is blind to whether it's evaluated or not, so it just introduces noise when comparing TASTs. "evaluated or not" isn't supposed to be very deterministic anyway.

So I just print lazy strings as their string values.

Reviewed By: madgen

Differential Revision: D50364803

fbshipit-source-id: 795dc35d1dd9114dc5ffa995402143defa3fcddd