typing for class<T>
Summary:
Implement `class<T>` in the typechecker as a new unary hint constructor `Hclass_args`. Currently, this hint decls a `Tapply` of `HH\classname` applied to its argument. The generality is necessary to be able to transform `classname<T>` to `class<T>`.

Long-term, the direction is to define an `Hclass of string` type that only represents the class name instead of an application, and migrate `Hclass_args` to `Hclass`, fixing all the odd typeholes in the process.

Reviewed By: viratyosin

Differential Revision: D49944968

fbshipit-source-id: e77e562e7be8ecb3f831c2d1617284754869bbda