Remove self argument to the trait forwarding method
Summary:
This diff removes the added `self` argument inside the body of the trait forwarding method.

When a static trait method `T$static.foo($this)` is defined,

* hackc adds the `self` argument to it, i.e. `T$static.foo($this, self)`,
* and it adds a forwarding method `T.foo($this, self)`,

then actual argument for `self` is given at analysis phase by Pulse.

The problem is that in the body of `T.foo`, it calls `T$static.foo` with two arguments:

```
define T.foo($this, self) {
  return T$static.foo($this, self);
}
```

But, it should be something like

```
define T.foo($this, self) {
  return T$static.foo($this); // self will be given by Pulse
}
```

Reviewed By: vsiles

Differential Revision: D50374096

fbshipit-source-id: 434161d6866558b373e2b56de401719ab968a848