Report an incompatible struct initializer as an error
Summary:
Report an incompatible struct initializer as an error. For example, invalid Thrift code such as

```
struct S {}

const S s = "";
```

would previously only give a warning and likely produce invalid generated code. With this change it gives the following error:

```
[ERROR:test.thrift:3] string is incompatible with `S`
```

This revealed bugs in the patch mutator which produced invalid Thrift AST with objects of struct types initialized with bools. These bugs are fixed as well.

And finally there was a fun bug in fixtures:

```
const SimpleStruct EMPTY = "{}";
```
which is fixed too.

Reviewed By: thedavekwon

Differential Revision: D50338734

fbshipit-source-id: a474214ebd9df2376e57b5c74fdabfa8445bf577