carry server init_id into CLIENT_CHECK.data
Summary:
We found that 2/3 of the time that users doing "hh" find it taking 2+ minutes, is because it had to init from saved-state.

nt591 and I wondered: is that because the saved-state-init was abnormally slow? To answer that, we'd need to correlate the CLIENT_CHECK representing the users "hh" invocation, with the SAVED_STATE_LOAD_OK event that the server must have done some time earlier.

This diff is the minimal way to achieve that. It adds the server's init_id to CLIENT_CHECK telemetry. That way, someone can JOIN this CLIENT_CHECK to the corresponding SAVED_STATE_LOAD_OK.

Differential Revision: D50458848

fbshipit-source-id: fbeebbeb608be058a65c4ecbe311743482320751