Create TaggedLowPtr type
Summary:
An augmented LowPtr with the bit layout
```
[...LOWPTR][TT]
```
where we take advantage of 4-byte alignment to stuff an enum into the low two bits.

The application here is to be able to discriminate ClassConstraints in union types, as they are currently assumed to all point to object types i.e. the
```
case type C = A | class<A>;
// same rep as
case type C = A | A;
```
This will be done in a follow up diff, the goal here is to just introduce the new type and replace `m_class`'s type with it.

Reviewed By: ricklavoie

Differential Revision: D50251849

fbshipit-source-id: 3609d342f6ac802bb08f6bdc733e70518a13aa8e