Fix nested key existence checking under coalesce
Summary:
We were only checking shape key existence at the top-level of a coalesce expression, this misses bugs where the inner shape doesn't have a key.

This diff fixes it by structurally recursing on the `exp['a']['b']['c']['d']... ?? ...` pattern.

The existing implementation is very ugly and we should instead fix `??` handling within the core typechecker. A TAST check is not suitable for this check due to the contextual effects of `??`.

Reviewed By: CatherineGasnier

Differential Revision: D50330486

fbshipit-source-id: 1d1c474190c2fa9c94e1193f2f98f1f4d130ca76