Generalise |> await support
Summary: Allow await to be used inside of other awaits, not just on the rhs of `|>`. For example, `await f(await g(), await h())`. This is still an UnstableFeature protected by the `pipe_await` file attribute. We allow nested awaits as long as they are a sequential composition of concurrent compositions (as in the above example). We signal a (parse) error if it requires a concurrent composition of sequential compositions. For exmaple `await f() + await g(await h())` is an error because we need to await `h()` before awaiting `g()`, and it's not clear which `f()` should be awaited alongside. hhvm doesn't support awaiting `f()` alongside the sequential composition.

Reviewed By: aorenste

Differential Revision: D50267974

fbshipit-source-id: c3a5414b90813a3864019c6702cae6a53882d37d