Attempt to fix duplicate flag issue (#9400)
Summary:

X-link: https://github.com/facebook/fbthrift/pull/579

CacheLib noticed OSS build failures due to `JSONProtocolCommon.cpp` being linked into the build twice, after we added a GFLAG to that file.

This diff attempts to fix the issue by depending on `thriftprotocol` rather than duplicating the files.

The code claims that it couldn't depend on `thriftprotocol` due to a circular dependency (added in D24720566), but Sandcastle and GitHub Actions is passing, so it doesn't seem to still be an issue.

Differential Revision: D50505357