Don't emit slow path when module boundary check is known
Summary:
This diff generally has no effect since simplify.cpp will also implify CallViolatesDeploymentBoundary checks when the caller is known. However, in cases when EvalSimplifyIR is off, like in certain tests, we still want to simplify the IR emitted.

For example, in hphp/test/quick/IncDecProp_nopt.php, module boundary checks are properly simplified and elided, but deployment boundary checks are not when enforcement is turned on, leading to an incorrect free

One thing that I don't understand as clearly is *why* not eliding module boundary checks causes tests that recycle the TC multiple times to fail; I think it might be because module boundary checks generate a new SSA block that isn't expected and assertion fails on free? (https://www.internalfb.com/intern/testinfra/diagnostics/3659174902070250.281475067279269.1697643520/) Will investigate this as a followup for both modules and packages, but properly simplifying is likely the correct fix either way.

Reviewed By: jtwarren

Differential Revision: D50506306

fbshipit-source-id: df94411b69d9232bda8eca58bdaa4b5cb97d8f2b