making the feature to prefer thread waiter over fiber waiter as an option
Summary:
As discovered in T152018412, some of the mutexes (TimedMutex) can be accessed by both fibers and
threads, and the existing feature to let thread waiter to steal the mutex from fiber waiter could
cause a sort of live lock issue where the thread waiters keep spinning for resources protected by
the mutex, preventing the fibers who can produce the resource from making progress. The motivation
behind the feature introduced in D4209155 was to resolve the deadlock that could happen when the
mutex is accessed at the same time from the main context of the fiber. However, this could break the
fairness and should be considered as an option for the use case where the main context is not used.

Reviewed By: dmm-fb

Differential Revision: D45931701

fbshipit-source-id: a292d76ce3cfac51fb6f34a58fc447c287c72d22