Improve typed local enforcement
Summary: In cases where we can't enforce a typed local's type before assigning to it, instead assign to a temporary and then assign to the typed local after enforcing the type on the temporary. Examples are array assignments `$x[e1] = e2` where it might be possible for `$x` to be enforced to a shape type, list assignments `list($x, $y) = e` and operator assignments `$x += e`.

Reviewed By: vsiles

Differential Revision: D49546355

fbshipit-source-id: 4629a38774c04c5ebb46846e3b78667abf6776b3