don't iterate through all packages when issuing soft warnings in repo mode
Summary: Upon a deployment boundary violation, we either issue a warning or raise an error depending on whether the symbol being accessed is in a soft package or not (https://www.internalfb.com/code/fbsource/[599cffef4da7]/fbcode/hphp/runtime/vm/runtime.cpp?lines=519%2C534). In repo mode, the only way we could've called into a soft package is when that package is already included in the active deployment, in which case there's no need to go through *all* deployments in order to determine the callee's softness – just check the actively deployed packages.

Reviewed By: jamesjwu

Differential Revision: D50038674

fbshipit-source-id: 2e33e1aee0e67249a287eab960af17504a199a88