Assume EventBase destructor is run in the EventBase thread
Summary:
The destructor runs logic that conceptually runs in the EventBase thread, so we need to make sure that `dcheckIsInEventBaseThread()` passes even in strict mode.

Setting up `loopThread_` also improves diagnostics in the case where the destructor is called while another thread is still driving the loop (which is a race).

Finally, we can drop `invokingLoop_`, as `loopThread_` already checks for re-entrancy, and `invokingLoop_` is actually racy.

Reviewed By: dmm-fb

Differential Revision: D50697500

fbshipit-source-id: 4a4846735f08ba8f798386a0ae0dd6841620aa64