noop: use identical function with correct name in Lsp_fmt
Summary:
Before:

Use `print_documentRename` to print workspace edits in textDocument/codeAction and codeAction/resolve responses

After:

Use `print_workspaceEdit`. This is a no-op, since `print_documentRename` is an alias for `print_workspaceEdit`.

The goal of this change is to make the code clearer and more maintainable: code actions have nothing to do with document renaming.

Reviewed By: ljw1004

Differential Revision: D50537280

fbshipit-source-id: 0f72f2862b2e8d02a8e2317e737d04c8c1dd9e20