Optimise mcrouter::getRoutingKey()
Summary:
We don't need to construct a `std::string` at all when there is no salt, we can retrieve the `folly::StringPiece`.

We can check for the salt and then split the behaviour.

Looking at the profiling, my gut feeling is that the vast majority of time there is no salt anyway.

Reviewed By: stuclar

Differential Revision: D50503304

fbshipit-source-id: 52e24e60d8d8408635872a4b9e4e3aabdf4c3e49