var_dump -> debugger_dump
Summary:
Context:

Users are used to typing `=$x` to log the value of `$x` in the debugger and expect to be able to use it in notebooks.

The current behavior I implemented is to rewrite to `var_dump`.

`debugger_dump`, recently added by arnabde03, is better than `var_dump` because it respects `toDebugDisplay` (important for privacy): with this diff, we use it

Reviewed By: arnabde03

Differential Revision: D50790712

fbshipit-source-id: 8b43b78e9720ef0370c921c6f33eaace6d274b15