Add CacheState destructor
Summary: When loading a shared object that includes ThreadId, it will register the CacheState object with AtFork, but never unregister it. If the shared object is unloaded, and loaded again, it will again be registered, throwing a std::invalid_argument>("at-fork: append: duplicate"); exception. Fix is to add a destructor so it is unregistered when the shared object is unloaded.

Reviewed By: Gownta

Differential Revision: D50452697

fbshipit-source-id: 2ec7c0da1d44fa865e238d858b9d80a050374762