ocaml cap results at max_results
Summary:
Meta:

We're getting too many results from glean. We pass in max_results=100, but glean is still sometimes giving back tens of thousands of results.
https://fburl.com/scuba/hh_server_events/7ojqf81i
 {F1136289021}

This diff is a bandaid. It truncates the list we get back from glean (and also augments telemetry so that in cases where we get back too many, we'll include the full angle query text to make it easier to repro).

It's possible that this "crazy high number of results" was breaking find-all-refs: https://fb.workplace.com/groups/342676619986174/posts/1390488798538279

We still need to do a proper fix

Reviewed By: mheiber

Differential Revision: D50841546

fbshipit-source-id: 51869ee8f9061c720d4fc4c105ab81e8fbfe7f83