don't crash hh_server upon VSCode cancellation
Summary:
When the user does Find-all-refs in the IDE, ClientLsp shells out to `hh --ide-find-refs-by-symbol ... <streamfile>`. If the user cancels the action then ClientLsp sends a SIGKILL and deletes the streamfile.

However, hh_server gets a fatal exception if the file is deleted while it still has further batches to do!

How frequent a problem is this? hh_server crashes with a fatal exception about 30 times a day, and it seems about 40% of crashes are because of this streamfile problem. https://fb.workplace.com/groups/342676619986174/posts/1391637921756700

This problem might have been masked in the past because of another find-refs bug, where even if the user cancelled their find-refs, it still sometimes took ~30s for the IDE to be responsive enough to delete the streamfile. But the preceding diffs in this stack have made the IDE more responsive in this case, so maybe the problem will become more severe.

I made the change in a fairly straightforward way. Previously the unhandled exception happened when a new multiworker started up and attempted (failed) to open the streamfile. I fixed it by making the multiworker detect ENOENT in this case, and return a Result.t using the Error case to indicate a cancellation request. (deleted streamfile in this case being taken as a request to cancel). Note that multiworkers aren't allowed to end in exceptions -- that's why I used Result.t

Reviewed By: nt591

Differential Revision: D50908406

fbshipit-source-id: fc42b6fa74f485bd9254e9e5000f17901af2d01d