cancel faster
Summary:
When you do find-all-refs and it's a demanding one (e.g. 60s to find 60k results)  and you click the Cancel button in VSCode, hh_server continues to work on the find-all-refs for 30s+. You can see it in the VSCode status bar.

What's happening is that ClientLsp has already sent SIGKILL to the `hh_client --ide-find-refs-by-symbol` process, and has already deleted the streaming file, but hh_server is in the middle of running multiworker. Each job ("batch") takes up to 30s when cold. And even though subsequent batches will decline to embark upon any work once they see that the streamfile has been deleted, there's no mechanism for the existing batches to give up. That's why we see it taking 30s.

I was also disturbed by the pathology addressed earlier in this stack D50894937, where the 30 worker processes were each exclusive-locking the streamfile for each file's worth of results they wanted to write into it. Given 60k results, with 30 processes all competing, it's not surprising that the client got starved when it tried to acquire a read-lock.

This diff addresses both issues.
1. It makes the workers check for cancellation (as indicated by unlinking the streamfile) after each single file they check, up to a maximum of once every 0.2s. (We'll still be vulnerable to some of the monster files that themselves take 30s, and I might address that in future by adding sub-file cancellation, but that's a way off).
2. It makes the workers acquire exclusive-locks at a slower rate. The heuristic I used: within the first 10 seconds of find-refs it makes sense to write results to the streamfile on a per-file basis when we're eager to get the first few results as soon as possible, but after 10 seconds then we'll be completing batches at a rate of several a second so it's fine to write results on a per-batch basis by this point. Also, within those first 10 seconds, if a worker has already written 20 results so far then it might as well hold off on the rest until the end of the batch.

For the coding idiom here, I used `raise Cancelled` as the exit mechanism when we act upon a cancellation request. This is the idiom used in C# where it's beautiful and popular. It's also nicely compositional -- when eventually I come to add sub-file cancellation into `Typing_tast.compute_tast` then that can be done with an exception too. (The alternative, representing cancellation with a Result.t, would require altering signatures all the way up the callstack). I'd have liked to use `raise Cancelled` as the way for a multiworker to exit, but alas they're not allowed to end with an exception.

Multiworkers actually have a separate cancellation mechanism too. It's strictly for external cancellation, where an external party can put something on a file-descriptor, and the multiworker orchestrator responds to that. I've never liked external cancellation; I believe cancellation should always be done by passing a "cancellationToken" parameter to a function. In Ocaml, Lwt used to support external cancellation but after several years the maintainer disavowed it in favor of request-based like cancellationToken. C# always uses cancellationToken. In this diff I used "does streamfile exist" as a form of cancellationToken.

This diff addresses the VSCode scenario where we have a streamfile and the user clicks Cancel. It doesn't address the CLI scenario where a user did "hh --rename" or "hh --find-refs" and then does Ctrl+C out of it -- in these cases, hh_server will simply continue the operation until it's finished. I did attempt another diff D50844981 which used Ctrl+C as the cancellationToken (more precisely, it pinged the hh_client to see if it was still around). But it didn't work out, and could only be performed on the main thread anyway, hence only on a per-batch basis. So I dropped it for now.

Reviewed By: nt591

Differential Revision: D50914422

fbshipit-source-id: 7492da1faf4db609b82142719394748d4379397a