avoid blocking clientLsp while find-refs is busy
Summary:
I observed that when a "busy" find-refs was onderway (i.e. one that produced tens of thousands of results, many per second), then Hack IDE would block -- (1) requests like "hover" would just not be fulfilled, (2) when I clicked to cancel the find-refs operation, it sometimes took 30s before receiving news that the shellout to "hh --ide-find-refs-by-symbol" had finished.

What happened was this. ClientLsp is running its main Lwt loop. In addition, it runs a second Lwt loop to read from the streaming-find-refs file and append items into an Lwt queue. The second Lwt loop does this:
1. "fstat", and observe that there were more bytes in the find-refs streaming file
2. a blocking lock+read+unlock to get those bytes
3. often, once it had done, "fstat" revealed yet more bytes and it did those

I observed (through adding a load of Hh_logger.log statements) that it often blocked for .8s to acquire the read-lock. The only other place that locks is inside the server process. It is not holding the lock for a long time!
```
    Sys_utils.with_lock fd Unix.F_LOCK ~f:(fun () ->
        let (_ : int) = Unix.lseek fd 0 Unix.SEEK_END in
        Sys_utils.write_non_intr fd bytes 0 (Bytes.length bytes);
        Unix.fsync fd)
```
So if clientLsp is taking so long to acquire the lock, the only explanation is that there's contention -- e.g. 30 worker processes, all finding plenty of results, all having their batches finish in about 1-2s, so lots of things queuing up to acquire the write lock and clientLsp has to patiently wait its turn.

Naturally with Lwt, while the second Lwt loop is doing synchronous blocking, the first Lwt loop is unable to proceed. (Lwt has single-threaded asynchrony, not multi-threaded).

This diff switches the secondary loop's lock acquisition to async via Lwt. Therefore, the secondary loop no longer blocks. Therefore the main Lwt loop can proceed as normal.

Reviewed By: nt591

Differential Revision: D50894937

fbshipit-source-id: 027620d1c8884a16647e25de233065f9f67016a2