Cleanup todo!() lints in hphp/hack/
Summary:
As of D50571479, we're linting against `todo!()` in checked-in code. Mark the existing call sites in hack with `#[allow(clippy::todo)]` so we don't increase the number of lints when editing existing code.

I've tried to keep the attributes as scoped as possible. Where possible, I've put the attribute on individual functions, but in a few cases I put the attribute on the whole file.

Reviewed By: nt591

Differential Revision: D50713453

fbshipit-source-id: ee7b9d584a7b19eef53a5adec4877a6c9cc9c14d