Add latch synchronization primitive to HPHP::coro
Summary:
Latch is a synchronization primitive which lets all coroutines block
until they all reach the latch, then release them all. For emulated
coroutines there's a bit of an issue because you cannot really "block"
without deadlocking. In that case, Latch is just a no-op, so must be
used with care.

Reviewed By: kavoor

Differential Revision: D50802820

fbshipit-source-id: 3377b5de0f6a5b9682fb24691c347825b4a20094