Handle "orphan" FuncInfo2s in HHBBC
Summary:
If a class is uninstantiable, it does not get a ClassInfo2. However,
we hang the FuncInfo2s for the methods of the class off the
ClassInfo2. For historical reasons, even if a class is uninstantiable
and does not have a ClassInfo2, we'll still analyze it. The analysis
will just be very pessimistic. The analysis allows for a class without
a ClassInfo, but it *does not* allow for a function without a
FuncInfo. So, we need some place to store FuncInfos for methods on
uninstantiable classes.

This adds a MethodsWithoutCInfo container for them. It's just a
wrapper around a vector of FuncInfos. When flattening classes we
create them if we detect a class as being uninstantiable, or if we
already know the class is uninstantiable.

Reviewed By: kavoor

Differential Revision: D50873343

fbshipit-source-id: d4e199662373609d82ab6cf9e4b2067384f3cbd4