Fix race condition in update_prop_initial_values
Summary:
update_prop_initial_values had a race condition which was
uncovered by removing the chunks in parallel. Both a 86pinit and a
86sinit can provide updated prop initial values. When analyzing
constants, these funcs will be analyzed independently and we'll
likewise process their results independently, potentially in multiple
threads. As a result, two threads might be simultaneously setting Attr
bits on properties and checking/setting the
"hasBadInitialPropValues". For the most part this is harmless, but it
can trigger assertions. This didn't happen before because the two
funcs were overwhelmingly likely to be chunked together and run on the
same thread.

Fix this with my old buddy, the hashed lock array idiom. It's rare to
have to update prop initial values, so contention should be very low.

Reviewed By: jano

Differential Revision: D50918614

fbshipit-source-id: c57b65dcf5927598bd8bfb876fd2aeaab5bcadc9