Add operations on ClassGraph, use them in init-types
Summary:
Add useful operations on ClassGraph, like doing subclass or could-be
operations, as well as union and intersection.

Move the "set bad initial prop values" calculation into the init-types
pass, as it really requires accurate type information (which we do not
have until then).

Add the notion of a "conservative" class, which is one whose subclass
list is too large to efficient represent. These classes will be dealt
with conservatively. Most of the machinery is there, but are not
actually generated yet. We cannot do so until we move ClassInfo to use
ClassGraph.

Reviewed By: kavoor

Differential Revision: D50814736

fbshipit-source-id: c9849beb88ed3e1a021fbfd93a3599355115b590