Remove coroutine emulation layer
Summary:
Remove the coroutine emulation layer and use coroutine syntax
directly. We had this emulation for compilers which did not support
coroutines fully, but it's not clear that this is needed
anymore. Furthermore, "emulating" coroutines was error prone, not well
tested, and it's unclear if it ever really worked. Remove it.

The only thing left in hphp/util/coro.h are some utility classes which
are not otherwise available.

Reviewed By: jano

Differential Revision: D50859518

fbshipit-source-id: 7740430d4def40dae5c48a1ff8f0457eabf7592d