Modify Hack Typechecker to Identify Weak Typing Callsites for FBIDs & IGIDs
Summary:
## Context
The idea behind this work-stream is to take strongly typed IGID_of<IMETAEntInstagramUser> and FBID_of<IMETAEntInstagramUser> and make sure that the types remain strong across the code-base. This means that we want to avoid converting them to ints or strings via casting, or through function parameter definitions.

## Diff Context
This diff modifies the Hack typechecker to identify locations where the code goes from being a strong type of one of these FBID/IGIDs to a weaker type

Reviewed By: madgen

Differential Revision: D50606768

fbshipit-source-id: 011237fa84a8979aad42b82cc777e28277fe4dfd