Mass replay: gather error hashes too
Summary:
In theory, some user changes might produce new typing errors without changing TASTs, for example:
i.php:
```
interface I {}
```
a.php:
```
class A implements I {}
```
f.php:
```
function f(A $x): void {
  expect_I($x);
}

function expect_I(I $_): void {}
```
Changing a.php to remove `implements I`:
```
class A {}
```
will produce an error in f without changing its TAST.

So to check fanouts and other change, we should not only compare TASTs but also errors.

Because it's hard to gather the full errors (the JSON format for variants produced by ocaml is not the same as the one produced in rust, pos_impl.rs does not exactly reflect pos.ml, etc.), we gather hashes of the errors as a set of hashes.

Start by reviewing tast_hashes.ml, I've highlighted the crux of the diff there.

There is a lot of boilerplate due to deriving yojson ppx. This is an annoying ppx which will throw warning 66 if you don't use one of the primitives like yojson_of_int in all of the files. I needed to derive yojson for ISet, which raised the warning for all files in `collections`, I could have separated iset.ml in its own library, but I decided to instead derive yojson for most types in `collections` and silent the warning in the remaining places. Suggestions welcome.

Differential Revision: D50837983

fbshipit-source-id: 87b1b5f6bedc773e1b5a707c0da5bb2f453bae55